refactor(angular): separates StorybookModuleMetadata from rendering service
and rename some function for more clarity
