+++ app/angular/src/client/preview/angular-beta/StorybookModule.test.ts
@@ -1,65 +1,11 @@
 import { Component, EventEmitter, Input, NgModule, Output, Type } from '@angular/core';
-import { platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { TestBed } from '@angular/core/testing';
 import { BehaviorSubject } from 'rxjs';
-import { RenderNgAppService } from './RenderNgAppService';
+import { getStorybookModuleMetadata } from './StorybookModule';
 
-jest.mock('@angular/platform-browser-dynamic');
-
-declare const document: Document;
-describe('RenderNgAppService', () => {
-  let renderNgAppService: RenderNgAppService;
-
-  beforeEach(async () => {
-    document.body.innerHTML = '<div id="root"></div>';
-    (platformBrowserDynamic as any).mockImplementation(platformBrowserDynamicTesting);
-    renderNgAppService = new RenderNgAppService();
-  });
-
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('should initialize', () => {
-    expect(renderNgAppService).toBeDefined();
-  });
-
-  describe('render', () => {
-    it('should add storybook-wrapper for story template', async () => {
-      await renderNgAppService.render({
-        storyFnAngular: {
-          template: '🦊',
-          props: {},
-        },
-        forced: false,
-        parameters: {} as any,
-      });
-
-      expect(document.body.getElementsByTagName('storybook-wrapper')[0].innerHTML).toBe('🦊');
-    });
-
-    it('should add storybook-wrapper for story component', async () => {
-      @Component({ selector: 'foo', template: '🦊' })
-      class FooComponent {}
-
-      await renderNgAppService.render({
-        storyFnAngular: {
-          props: {},
-        },
-        forced: false,
-        parameters: {
-          component: FooComponent,
-        },
-      });
-
-      expect(document.body.getElementsByTagName('storybook-wrapper')[0].innerHTML).toBe(
-        '<foo>🦊</foo>'
-      );
-    });
-  });
-  describe('getNgModuleMetadata', () => {
+describe('StorybookModule', () => {
+  describe('getStorybookModuleMetadata', () => {
     describe('with simple component', () => {
       @Component({
         selector: 'foo',
@@ -98,7 +44,7 @@ describe('RenderNgAppService', () => {
           localFunction: () => 'localFunction',
         };
 
-        const ngModule = RenderNgAppService.getNgModuleMetadata(
+        const ngModule = getStorybookModuleMetadata(
           { storyFnAngular: { props }, parameters: { component: FooComponent } },
           new BehaviorSubject(props)
         );
@@ -130,7 +76,7 @@ describe('RenderNgAppService', () => {
           },
         };
 
-        const ngModule = RenderNgAppService.getNgModuleMetadata(
+        const ngModule = getStorybookModuleMetadata(
           { storyFnAngular: { props }, parameters: { component: FooComponent } },
           new BehaviorSubject(props)
         );
@@ -151,7 +97,7 @@ describe('RenderNgAppService', () => {
         };
         const storyProps$ = new BehaviorSubject(initialProps);
 
-        const ngModule = RenderNgAppService.getNgModuleMetadata(
+        const ngModule = getStorybookModuleMetadata(
           { storyFnAngular: { props: initialProps }, parameters: { component: FooComponent } },
           storyProps$
         );
@@ -199,7 +145,7 @@ describe('RenderNgAppService', () => {
         };
         const storyProps$ = new BehaviorSubject(initialProps);
 
-        const ngModule = RenderNgAppService.getNgModuleMetadata(
+        const ngModule = getStorybookModuleMetadata(
           { storyFnAngular: { props: initialProps }, parameters: { component: FooComponent } },
           storyProps$
         );
