+++ app/angular/src/client/preview/angular-beta/utils/NgComponentAnalyzer.test.ts
